007-01: Add field Handle to Voice Class, change arguments in voice container add and remove to uint8_t, added nonstatic method doSend to access non static members of class OscService, just as explained in stratify

+++ lib/VoiceContainer/src/VoiceContainer.h
@@ -19,9 +19,9 @@ class VoiceContainer : public Module {
 public:
     VoiceContainer();
 
-    void add(int Handle);
+    void add(uint8_t Handle);
 
-    void remove(int Handle);
+    void remove(uint8_t Handle);
 
     static void receive(void *context, OscMsgChild &message, uint8_t initialOffset);
